feat(auth): implement login flow and update JWT secret handling 	- Added authentication routes and handlers 	- Implemented auth_service for login logic 	- Updated JWT secret key to use []byte type 	- Integrated auth_handler into router 	- Minor updates in user and password packages

+++ internal/handler/user_handler.go
@@ -96,7 +96,7 @@ func (handler *UserHandler) Update(c *gin.Context) {
 		return
 	}
 
-	user.ID = uint(userID)
+	user.ID = userID
 
 	// Call service layer to create user
 	user, err := handler.userService.Update(user)
